css-snap-size-1/snap-width-parsing-001.html

<!DOCTYPE html>
<title>CSS Snap Size: parsing snap-width</title>
<link rel="author" title="Koji Ishii" href="kojiishi@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-snap-size/#snap-width">
<meta name="assert" content="This test asserts the parsing and getComputedStyle works correctly for the snap-width property.">
<meta name="flags" content="dom">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div data-expected="0px" title="initial value"></div>

<div style="snap-width: 20px" data-expected="20px">
  <div data-expected="0px" title="snap-width should not inherit"></div>
  <div style="snap-width: inherit" data-expected="20px"></div>
</div>

<div style="snap-width: 20px; snap-width: initial" data-expected="0px"></div>

<!-- "0" is a valid length even without units -->
<div style="snap-width: 20px; snap-width: 0" data-expected="0px"></div>

<!-- These are invalid and thus its preceding declarations should remain -->
<div style="snap-width: 20px; snap-width: 1" data-expected="20px"></div>
<div style="snap-width: 20px; snap-width: -1px" data-expected="20px"></div>
<div style="snap-width: 20px; snap-width: 10%" data-expected="20px"></div>

<div id="pt" style="snap-width: 40pt"></div>

<script>
Array.prototype.forEach.call(document.querySelectorAll("[data-expected]"), function (element) {
  test(function () {
    var actual = getComputedStyle(element).snapWidth;
    assert_equals(actual, element.dataset.expected);
  }, element.title || element.getAttribute("style"));
});

test(function () {
  assert_equals(getComputedStyle(document.getElementById("pt")).snapWidth.slice(-2), "px");
}, "Computed value must be the absolute length");
</script>
